+++ geoopt/samplers/sgrhmc.py
@@ -3,7 +3,6 @@ import math
 import torch
 
 from geoopt.tensor import ManifoldParameter, ManifoldTensor
-from geoopt.manifolds import R
 from geoopt.samplers.base import Sampler
 from ..utils import copy_or_set_
 
@@ -66,7 +65,7 @@ class SGRHMC(Sampler):
                         if isinstance(p, (ManifoldParameter, ManifoldTensor)):
                             manifold = p.manifold
                         else:
-                            manifold = R()
+                            manifold = self._default_manifold
 
                         egrad2rgrad = manifold.egrad2rgrad
                         retr_transp = manifold.retr_transp