+++ geoopt/tensor.py
@@ -1,5 +1,5 @@
 import torch.nn
-from .manifolds import R, Manifold
+from .manifolds import Euclidean, Manifold
 from .docutils import insert_docs
 from .utils import copy_or_set_
 
@@ -15,7 +15,7 @@ class ManifoldTensor(torch.Tensor):
         A manifold for the tensor, (default: :class:`geoopt.R`)
     """
 
-    def __new__(cls, *args, manifold=R(), requires_grad=False, **kwargs):
+    def __new__(cls, *args, manifold=Euclidean(), requires_grad=False, **kwargs):
         if len(args) == 1 and isinstance(args[0], torch.Tensor):
             data = args[0].data
         else: